move fav edition to sidebar
